Rename `shares` directory
