+++ src/forms/form.share.spec.ts
@@ -2,9 +2,9 @@ import { inFormElement, inGroup } from '@frontmeans/input-aspects';
 import { ContextBuilder } from '@proc7ts/context-values';
 import { AfterEvent } from '@proc7ts/fun-events';
 import { BootstrapContext, Component, ComponentContext, ComponentSlot, DefinitionContext } from '@wesib/wesib';
-import { ComponentShare, ComponentShare__symbol } from '../share';
-import { ComponentShareRegistry } from '../share/component-share-registry.impl';
-import { SharedByComponent$ContextBuilder } from '../share/shared-by-component.impl';
+import { ComponentShare, ComponentShare__symbol } from '../shares';
+import { ComponentShareRegistry } from '../shares/component-share-registry.impl';
+import { SharedByComponent$ContextBuilder } from '../shares/shared-by-component.impl';
 import { testDefinition, testElement } from '../spec/test-element';
 import { FieldShare } from './field.share';
 import { Form } from './form';
