+++ src/index.ts
@@ -6,6 +6,6 @@ export * from './automount';
 export * from './fetch';
 export * from './hierarchy';
 export * from './navigation';
-export * from './share';
+export * from './shares';
 export * from './tree';
 export * from './util';
