+++ src/shares/component-share-key.impl.ts
