+++ src/shares/component-share-locator.ts
