+++ src/shares/component-share-ref.ts
