+++ src/shares/component-share-registry.impl.ts
