+++ src/shares/component-share.impl.ts
