+++ src/shares/component-share.ts
