+++ src/shares/component-shareable.ts
