+++ src/shares/share-accessor.impl.ts
