+++ src/shares/target-component-share.ts
